requirements.txt
anthropic
datasets
google-genai
numpy
matplotlib
openai
pillow
seaborn
torch
transformers
tqdm
